Update NavGraph.kt
I removed categories screen in the bottom nav bar since i changed my idea and i wanted to add a discovery screen for the ease of having all the categories in the same screen. And i made my products to show their own product details screen based on their ID's, but due to the time left, i changed it to direct each and every product to to the same product details screen when clicked.